Add -Wno-error=format-nonliteral to allow gnc-backend-sql.c to compile
gnc-backend-sql.c uses a format string from a GncSqlBackend that
is set based on the type of SQL database that is targeted.

At least when compiling with clang, this triggers -Wformat-nonliteral
which -Werror turns into a compiler error.

This patch makes a targeted change in src/backend/sql/Makefile.am
for -Wno-error=format-nonliteral in this scope only