Bug 747377: Fix overly restrictive input validation for IBAN of SEPA transfer.
Only in some countries the IBAN is really restricted to numeric-only
(most notably in Germany, "DE"). In some other countries parts of the
IBAN may be alphas. These checks could be extended for more countries,
but since aqbanking with the SEPA transfers is used mostly in Germany,
we just leave it with the exception rule for Germany.