Abstract options handling to QofBook.
Eliminates direct KVP access in app-utils, all of which centered around
options.

Beneficial side effect: With all option writes now handled by
qof_book_set_option(), the problem of options not being committed is
eliminated.

The options system is unnecessarily complex, with much back-and-forth
between C and Guile. It needs to be completely rewritten, but that's a
project for its own branch.