Bug 752204 - .gml2 files are modified during build, take 2.
Turns out that there are changes as well as compression, particularly
if the timezone is different from the one from the last commit of the
gml2 files. Since there's no real benefit to saving the files most of
the time, don't, but leave the save line in, commented out, to make it
easy to write new files if necessary.