Make backend sync errors survive to the session.
The backends were using qof_backend_get_error() to test for sync errors.
This function clears the error, so the tests resulted in the error being
cleared before the session could see it and so it thought that the sync
had succeeded.

Replace those uses of qof_backend_get_error() with a new function
qof_backend_check_error() that doesn't clear the error.