Define build-report! function instead of loop.
The existing code uses tuples to define the properties of the individual
reports created.  Replace that by a function to build one of them which
gets the properties passed as arguments, and use a series of calls to
replace the old loop.  This makes the properties much more easier to
read.