Bug 608098 - Option is hard to find: Change Transaction Report to show account names in multirow txn
The original commit replaces the General->Style option with
a similar Display->Detail Level option.
This change is done because some columns can or can't be displayed
depending on this option. Listing these together makes this much
more clearer.

This text should go into the release notes:
On the Transaction report the General->Style option has been
replaced with a similar Display->Detail Level option. This
was done in order to make the display of some columns dependent
on the value of this option.

When upgrading from 2.6.x or earlier to 2.8.x and you had saved
or open reports that had set General->Style to 'Multi-line'
you will have to update these reports to set 'Multi-line' on the
Display->Detail Level option instead and optionally resave your report.