Don't require prices in Scheduled Transactions with marker commodity splits.
Bug 754192 - Since Last Run Dialog asks for security price even if not needed
Bug 662126 - Cannot create scheduled transaction for dividend reinvestment

If a split in an SX has a 0 amount, don't make its commodity the transaction
commodity, set the multi-commodity flag, or create a variable for it.

Make the warning about not being able to auto-create SXes with variables
or multiple commodities more explicit about the latter.