Optimize pricedb so it works better with really big price lists.
Opening my accounts file took over 10 minutes before this change, most of
it spent populating the account balances in account tree, and most of that
spent in g_sort_list.  Expanding or collapsing subtrees in the account tree
also took several seconds.  This change reduces the use of g_sort_list and
reduces the length of lists sorted when it is used.  Opening the file is
now only a few seconds slower than before the PriceDB changes.