Set val_imbalance in gnc_transaction_balance_trading.
When the code was refactored in 61973a8 the return value from the new
function gnc_transaction_get_commodity_imbalance was ignored.  This had
the effect of using an uninitialized variable to set the value in a new
trading account split that is not for the transaction's currency.  That
means that multicommodity transactions created with this bug and trading
accounts on will not be balanced.