Bug 745101 - No warning when quitting with unsaved file
Improved fix: mark session dirty in qof_commit_edit_part2
instead of delegating this to the backends. This way
a session is also properly marked dirty (and hence the
save button enabled) when no backend is chosen yet
(that is, when in a new file that's never saved yet).

For the xml backend this means there's nothing to do anymore
at commit time, so the xml_commit_edit callback is removed.