Use @SHELL@ instead of hardcode /bin/sh in test shell scripts.
This is a workaround for MacOS X 10.11's System Integrity Protection,
aka SIP, which among other things prevents system binaries from passing
linker environment variables to children. Anything in /bin or /usr/bin
qualifies as a system binary, and that includes /bin/sh and /bin/bash.
Not being able to pass linker paths breaks all of our guile tests which
need to tell Guile where to look for the libraries being tested.

To apply the workaround one needs to first have a shell outside of /bin
and /usr/bin. Simply copying /bin/sh to your home direcory works, as
does building a suitable shell somewhere. Just pass the path to that
shell in $CONFIG_SHELL to configure and all of the shell scripts and
libtool .la files will be created with it in the shebang line and the
linker environment variables will be passed in.

In order to build GnuCash in the first place you'll also need to edit
guild to use the alternate shell.