Bug 756335 - When importing, date selection causes exit crash
Corrects 2f2d1d9 and reverts 7aa9b98 to correctly prevent the second
crash from the bug. Also moves the re-parse of the dates from
gnc_ui_qif_import_date_valid_cb to gnc_ui_qif_import_account_prepare so
that it happens after the "forward" button on the date format page is
clicked instead of after the first time the combo box is used, allowing
the user to change her mind and have it work.