Bug 728722 - Setting number format details appear wrong in Help, section 10.3.4. Counters Book Options Tab
This is a follow-up commit to fix the core of the issue.
With this commit gnucash is more liberal at accepting
counter formats. It will accept either li, lli, I64i and
whatever is defined for G_GINT_64 or PRIx64 on the user's
platform. Internally the code will always convert the
specifier set by the user with PRIx64, which should always
be the correct one on any platform.

Additionally a few extra tests were added to stress the
counter format code a bit more.