Bug763257 - add line charts to Income/Expense/Asset/Liability reports
Fix date formatting for UK and locale choices.
The fix works around what's most likely a bug in jqplot
that it poorly interprets UK formatted date strings. To avoid
the bug this patch will always pass ISO formatted date
strings to jqplot for line charts. This is only applied to
the liability chart for now and should be done for all other
line charts as well.