When default tax table is 'None', set a new invoice/bill entry as not taxabl by default as well.
This slightly changes the semantics of a 'None' default tax table
from 'taxable but no tax details given'
to 'not taxable'.
While not exactly the same the former is a highly unlikely
situation, where the latter is more common and hence makes
more sense. In the rare case a user does want to express
the old semantics this is still possible by explicitly
creating a tax table without any details.
Note this only influences newly entered entries. It has
no influence on existing ones.