Bug 137017 - date of transaction change with time zone change
First step: Change the timestamp to 11:00 UTC instead of midnight local,
adjusting by an hour or two if the local timezone is one near the
International Date Line to keep the date from flipping around.

Scrub all old entries to make current files correct.

Note: This effectively disposes of the distinction of close-book transactions
having a noon instead of midnight timestamp as a way to distinguish them
from regular transactions, but that distinction doesn't seem to be used;
xaccTransIsCloseBook() is used instead.