Bug 756416 - CSV data with Posted information will always have last invoice not posted.
When importing the last line of the CSV the gtk_tree_model_get() never happens
and the following compare uses the previous value of new_id, which is now = id,
so the last invoice is never posted.

Set new_id = Null before gtk_tree_model_get() to ensure processing the last line.