Change associate file/link to show existing entry
Currently there is no way of knowing what the existing association is
and so what dialog to use. Changed the file dialog to use existing uri
along with a label in case the uri does not exist and populated the
entry in the link dialog with the existing uri.

Also added a test for when entry is cleared and also being NULL to
control menu sensitivity and also added a remove button to clear entry