Add check for path being relative.
If the preference "assoc-head" is set and file path is a sub directory
then only this part is saved to the transaction making the full path
portable. Paths outside of "assoc-head" will be saved in full. There is
also a test to make sure the URI passed to the open call has a valid
scheme otherwise the windows version will crash.