Bug 770303 - Incorrect commodity match during import OFX and possible solution.
Comparing only on the strlen of the imported commodity produces false
matches when the account commodity string begins with the same
characters but is longer. E.g. if the imported commodity is F (Ford
Motor) and Facebook (FB) is encountered first in the account list then
the matcher will select Facebook, but that will fail when the commodity
GUIDs are compared.