Created option to have a pixel and percent spinner
Created an option for plot sizes so you have a spinner for pixels and
one for percentage on the same line. The value set and returned are
positive for Pixel values and negative for Percentages. This should be
a drop in replacement for any use of the existing plot size option.