Prevent the user from setting an invalid settings name
The keyfile that stores the settings won't accept '[' and ']' as settingss names,
so this commit will prevent the user from entering these characters. They are
automatically replaced with the valid '(' and ')' characters.

In addition any attempt to save settings with an invalid name via a code path
bypassing the gui will be refused.