Travis - fix failure to load some libraries while running scm tests
In the travis environment (guile 1.8 in combination with other unidentified,
particular package versions) the scm based tests are looking for
backend modules (xml and dbi) in the installation prefix instead of
in the build directory. Explicitly setting the environment variables
GNC_BUILDDIR and GNC_UNINSTALLED fix this. I haven't figured out why this
is not needed on all platforms/tool version combinations. But the fix
works in all (autotools based) cases. CMake builds may need a similar
fix, but I don't know that environment sufficiently to implement it.