Drop callback that reloads report when a report tab is selected
This has the unwanted side-effect of misaligning the chart
legends (due to some internal bug in webkit). Resizing the window
properly redraws the chart afterwards, but leaving the page and
returning reloads the page again, hitting the bug again. This
prevents multiple report tabs open at once to show a correct chart.
Each resize of the window to correct one chart will trigger a
misaligned legend on all other charts the next time you look at them.

The disadvantage of removing this code is that reports loaded
automatically when gnucash opens are not sized properly. (All but
the one in the very last tab of the book being opened, that is).
However these pages can be corrected by manually reloading them
once. That's annoying but at least allows for correctly displayed reports.

A proper solution must be found for this of course.