Add function to clear the display of the currency edit widget
When first created, the display is blank and in this position will
return the default currency. If a currency is selected, there is no way
to go back to this state. The only way to do this is destroy the widget
and create a new one so added a function to do this without destroying
the widget. This is used in the next commit.