Change the cell grid lines slightly
Grid lines will be generally be drawn by subtracting 0.5 from the x and
y values apart from the left border of the first column cell has 0.5
added to the x value and the first row cell has 0.5 added to the top
border y value. By doing this the overall dimensions does not change
and all borders are visible.