Remove the gncmod boilerplate from standard-reports and utility-reports
This adds unnecessary overhead and also serves as an initial experiment
to greatly reduce the gnc-module based modularization.

Oh, and it fixes the autotools make check failure in standard-reports
which I introduced in the previous commit. The fix was to either
fully gnc-modularize business reports or to go to de-gnc-modularize
standard-reports. I chose the latter as I don't want to add any new
gnc-modules any more at this point.